requirements.txt
# EvidenceIQ - Production-Grade Local Multimodal Intelligence Platform
# Zero cloud transmission. Privacy-first. RBAC-enforced.

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27

# Configuration
pydantic-settings==2.6.1

# Image Processing
pillow==11.0.0
opencv-python==4.10.0.84
piexif==1.1.3

# Document Parsing
pymupdf==1.24.14

# Local AI (Ollama)
ollama==0.4.0

# Vector Database & Embeddings
chromadb==0.5.20
sentence-transformers==3.3.0
transformers==4.46.3
torch==2.5.1

# PII Detection (local, no cloud)
presidio-analyzer==2.2.355

# PDF Generation
reportlab==4.2.5

# Logging
structlog==24.4.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
